+++ IBAnimatable/AnimatableNavigationController.swift
@@ -11,24 +11,26 @@ public class AnimatableNavigationController: UINavigationController, TransitionA
   @IBInspectable public var transitionAnimationType: String?
   @IBInspectable public var transitionDuration: Double = .NaN
 
+  // MARK: - Lifecylce
   public override func viewDidLoad() {
     super.viewDidLoad()
-    configureNavigationControllerDelegate(transitionAnimationType, transitionDuration: transitionDuration)
+    configureNavigationControllerDelegate()
   }
 
+  // MARK: - Private
   // Must have a property to keep the reference alive because `UINavigationController.delegate` is `weak`
   private var navigator: Navigator?
 
-  private func configureNavigationControllerDelegate(transitionAnimationType: String?, transitionDuration: Duration) {
-    guard let transitionAnimationType = transitionAnimationType else {
+  private func configureNavigationControllerDelegate() {
+    guard let transitionAnimationType = transitionAnimationType, animationType = TransitionAnimationType(rawValue: transitionAnimationType) else {
       return
     }
     var duration = transitionDuration
     // Set the default duration for transition
     if transitionDuration.isNaN {
-      duration = 0.5
+      duration = 0.35
     }
-    navigator = Navigator(transitionAnimationType: transitionAnimationType, transitionDuration: duration)
+    navigator = Navigator(transitionAnimationType: animationType, transitionDuration: duration)
     delegate = navigator
   }
 }