+++ IBAnimatable/AnimatedTransitioning.swift
@@ -10,9 +10,9 @@ import UIKit
 public protocol AnimatedTransitioning: UIViewControllerAnimatedTransitioning {
   
   /**
-   String value of `TransitionAnimationType` enum
+   Value of `TransitionAnimationType` enum
    */
-  var transitionAnimationType: String { get set }
+  var transitionAnimationType: TransitionAnimationType { get set }
   
   /**
    Transition duration
@@ -22,15 +22,26 @@ public protocol AnimatedTransitioning: UIViewControllerAnimatedTransitioning {
   /**
    Reverse animation type: used to specify the revers animation for pop or dismiss.
    */
-  var reverseAnimationType: String? { get set }
+  var reverseAnimationType: TransitionAnimationType? { get set }
 }
 
 public extension AnimatedTransitioning {
   public func retrieveViews(transitionContext: UIViewControllerContextTransitioning) -> (UIView?, UIView?, UIView?) {
     return (transitionContext.viewForKey(UITransitionContextFromViewKey), transitionContext.viewForKey(UITransitionContextToViewKey), transitionContext.containerView())
   }
-
-  public func animateWithCATransition(transitionContext: UIViewControllerContextTransitioning, type: CATransitionType, subtype: String?) {
+  
+  public func retrieveViewControllers(transitionContext: UIViewControllerContextTransitioning) -> (UIViewController?, UIViewController?, UIView?) {
+    return (transitionContext.viewControllerForKey(UITransitionContextFromViewControllerKey), transitionContext.viewControllerForKey(UITransitionContextToViewControllerKey), transitionContext.containerView())
+  }
+  
+  public func retrieveTransitionDuration(transitionContext: UIViewControllerContextTransitioning?) -> NSTimeInterval {
+    if let transitionContext = transitionContext {
+      return transitionContext.isAnimated() ? transitionDuration : 0
+    }
+    return 0
+  }
+  
+  public func animateWithCATransition(transitionContext: UIViewControllerContextTransitioning, type: SystemTransitionType, subtype: String?) {
     let (_, tempToView, tempContainerView) = retrieveViews(transitionContext)
     guard let toView = tempToView, containerView = tempContainerView else {
       transitionContext.completeTransition(true)
@@ -40,15 +51,16 @@ public extension AnimatedTransitioning {
     containerView.addSubview(toView)
     CALayer.animate({
       let transition = CATransition()
-      transition.type = String(type)
+      transition.type = type.rawValue
       if let subtype = subtype {
         transition.subtype = subtype
       }
       transition.duration = self.transitionDuration(transitionContext)
-
+      
       containerView.layer.addAnimation(transition, forKey: kCATransition)
-    }) {
+    },
+    completion: {
       transitionContext.completeTransition(!transitionContext.transitionWasCancelled())
-    }
+    })
   }
 }