+++ IBAnimatable/TransitionFadeType.swift
@@ -1,0 +1,15 @@
+//
+//  Created by Jake Lin on 2/27/16.
+//  Copyright © 2016 Jake Lin. All rights reserved.
+//
+
+import Foundation
+
+/**
+ Transition Fade Type
+ */
+public enum TransitionFadeType {
+  case Fade     // ToView fades in and FromeView fades out
+  case FadeIn   // ToView fades in
+  case FadeOut  // FromView Fades out
+}